# frontend/openapi.regulus.json
{
  "components": {
    "schemas": {
...
      "CapabilityRankingRow": {
        "properties": {
          "aer": {
            "title": "Aer",
            "type": "number"
          },
          "capability_id": {
            "title": "Capability Id",
            "type": "string"
          },
          "capability_type": {
            "title": "Capability Type",
            "type": "string"
          },
          "is_protected": {
            "title": "Is Protected",
            "type": "boolean"
          },
          "net_margin_usd": {
            "title": "Net Margin Usd",
            "type": "number"
          }
        },
        "required": [
          "capability_id",
          "capability_type",
          "net_margin_usd",
          "aer",
          "is_protected"
        ],
        "title": "CapabilityRankingRow",
        "type": "object"
      },
      "CapabilityValueRow": {
        "properties": {
          "capability_id": {
            "title": "Capability Id",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "net_margin_usd": {
            "title": "Net Margin Usd",
            "type": "number"
          }
        },
        "required": [
          "capability_id",
          "net_margin_usd",
          "confidence"
        ],
        "title": "CapabilityValueRow",
        "type": "object"
      },
...
      "ConfidenceGateStatus": {
        "properties": {
          "blocked": {
            "title": "Blocked",
            "type": "integer"
          },
          "passed": {
            "title": "Passed",
            "type": "integer"
          }
        },
        "required": [
          "passed",
          "blocked"
        ],
        "title": "ConfidenceGateStatus",
        "type": "object"
      },
...
      "DataQualityMix": {
        "properties": {
          "inferred": {
            "title": "Inferred",
            "type": "integer"
          },
          "measured": {
            "title": "Measured",
            "type": "integer"
          },
          "mixed": {
            "title": "Mixed",
            "type": "integer"
          }
        },
        "required": [
          "measured",
          "inferred",
          "mixed"
        ],
        "title": "DataQualityMix",
        "type": "object"
      },
...
      "GroundTruthCostIn": {
        "properties": {
          "amount_usd": {
            "title": "Amount Usd",
            "type": "number"
          },
          "capability_id": {
            "title": "Capability Id",
            "type": "string"
          },
          "component": {
            "title": "Component",
            "type": "string"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          }
        },
        "required": [
          "period_start",
          "period_end",
          "capability_id",
          "component",
          "amount_usd"
        ],
        "title": "GroundTruthCostIn",
        "type": "object"
      },
      "GroundTruthImportRequest": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/GroundTruthCostIn"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "rows"
        ],
        "title": "GroundTruthImportRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImplementationCompareRow": {
        "properties": {
          "arm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arm"
          },
          "capability_id": {
            "title": "Capability Id",
            "type": "string"
          },
          "confidence_breakdown": {
            "additionalProperties": true,
            "title": "Confidence Breakdown",
            "type": "object"
          },
          "confidence_high_usd": {
            "title": "Confidence High Usd",
            "type": "number"
          },
          "confidence_low_usd": {
            "title": "Confidence Low Usd",
            "type": "number"
          },
          "implementation_id": {
            "title": "Implementation Id",
            "type": "string"
          },
          "net_margin_usd": {
            "title": "Net Margin Usd",
            "type": "number"
          }
        },
        "required": [
          "capability_id",
          "implementation_id",
          "net_margin_usd",
          "confidence_low_usd",
          "confidence_high_usd"
        ],
        "title": "ImplementationCompareRow",
        "type": "object"
      },
...
      "KPIResponse": {
        "properties": {
          "efficiency_index": {
            "title": "Efficiency Index",
            "type": "number"
          },
          "net_ai_margin_usd": {
            "title": "Net Ai Margin Usd",
            "type": "number"
          },
          "portfolio_confidence_score": {
            "title": "Portfolio Confidence Score",
            "type": "number"
          },
          "total_ai_spend_usd": {
            "title": "Total Ai Spend Usd",
            "type": "number"
          },
          "total_ai_value_usd": {
            "title": "Total Ai Value Usd",
            "type": "number"
          }
        },
        "required": [
          "total_ai_spend_usd",
          "total_ai_value_usd",
          "net_ai_margin_usd",
          "portfolio_confidence_score",
          "efficiency_index"
        ],
        "title": "KPIResponse",
        "type": "object"
      },
...
      "PolicyTimelineRow": {
        "properties": {
          "action_type": {
            "title": "Action Type",
            "type": "string"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "capability_id": {
            "title": "Capability Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "proposed_at": {
            "title": "Proposed At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "capability_id",
          "action_type",
          "status",
          "proposed_at"
        ],
        "title": "PolicyTimelineRow",
        "type": "object"
      },
...
      "TrendPoint": {
        "properties": {
          "x": {
            "title": "X",
            "type": "string"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "TrendPoint",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "AI Economic Control Plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/dashboard/action-suppression": {
      "get": {
        "operationId": "get_action_suppression_v1_dashboard_action_suppression_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrendPoint"
                  },
                  "title": "Response Get Action Suppression V1 Dashboard Action Suppression Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Action Suppression",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/calibration-trend": {
      "get": {
        "operationId": "get_calibration_trend_v1_dashboard_calibration_trend_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrendPoint"
                  },
                  "title": "Response Get Calibration Trend V1 Dashboard Calibration Trend Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Calibration Trend",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/capability-ranking": {
      "get": {
        "operationId": "get_capability_ranking_v1_dashboard_capability_ranking_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CapabilityRankingRow"
                  },
                  "title": "Response Get Capability Ranking V1 Dashboard Capability Ranking Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Capability Ranking",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/capital-destroyers": {
      "get": {
        "operationId": "get_destroyers_v1_dashboard_capital_destroyers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CapabilityValueRow"
                  },
                  "title": "Response Get Destroyers V1 Dashboard Capital Destroyers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Destroyers",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/confidence-gate-status": {
      "get": {
        "operationId": "get_conf_gate_status_v1_dashboard_confidence_gate_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfidenceGateStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Conf Gate Status",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/confidence-trend": {
      "get": {
        "operationId": "get_conf_trend_v1_dashboard_confidence_trend_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrendPoint"
                  },
                  "title": "Response Get Conf Trend V1 Dashboard Confidence Trend Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Conf Trend",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/data-quality-mix": {
      "get": {
        "operationId": "get_data_quality_mix_v1_dashboard_data_quality_mix_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataQualityMix"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Data Quality Mix",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/drift-timeline/{capability_id}": {
      "get": {
        "operationId": "get_drift_timeline_v1_dashboard_drift_timeline__capability_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "capability_id",
            "required": true,
            "schema": {
              "title": "Capability Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrendPoint"
                  },
                  "title": "Response Get Drift Timeline V1 Dashboard Drift Timeline  Capability Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Drift Timeline",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/efficiency-trend": {
      "get": {
        "operationId": "get_eff_trend_v1_dashboard_efficiency_trend_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrendPoint"
                  },
                  "title": "Response Get Eff Trend V1 Dashboard Efficiency Trend Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Eff Trend",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/estimated-vs-ground-truth": {
      "get": {
        "operationId": "get_estimated_vs_ground_truth_v1_dashboard_estimated_vs_ground_truth_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Estimated Vs Ground Truth V1 Dashboard Estimated Vs Ground Truth Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Estimated Vs Ground Truth",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/implementation-compare/{capability_id}": {
      "get": {
        "operationId": "get_implementation_compare_v1_dashboard_implementation_compare__capability_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "capability_id",
            "required": true,
            "schema": {
              "title": "Capability Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImplementationCompareRow"
                  },
                  "title": "Response Get Implementation Compare V1 Dashboard Implementation Compare  Capability Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Implementation Compare",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/kpis": {
      "get": {
        "operationId": "get_kpis_v1_dashboard_kpis_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KPIResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Kpis",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/policy-timeline": {
      "get": {
        "operationId": "get_policy_timeline_v1_dashboard_policy_timeline_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PolicyTimelineRow"
                  },
                  "title": "Response Get Policy Timeline V1 Dashboard Policy Timeline Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Policy Timeline",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/v1/dashboard/top-creators": {
      "get": {
        "operationId": "get_top_creators_v1_dashboard_top_creators_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CapabilityValueRow"
                  },
                  "title": "Response Get Top Creators V1 Dashboard Top Creators Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Top Creators",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/v1/reconciliation/ground-truth-import": {
      "post": {
        "operationId": "import_ground_truth_v1_reconciliation_ground_truth_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroundTruthImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Import Ground Truth V1 Reconciliation Ground Truth Import Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Ground Truth",
        "tags": [
          "reconciliation"
        ]
      }
    },
...
  }
}